add animation.css in index.css remove Keyframes.css move DavitNotification.tsx in molecules add DavitNotification.css
